fix a bug causing stage 2 bootlaoder truncated

+++ boot/loader/include/terminal.h
@@ -8,7 +8,7 @@
 
 void InitializeTerminal();
 
-void TerminalSetColor(WORD fgcolor, WORD bgcolor);
+void TerminalSetColor(DWORD fgcolor, DWORD bgcolor);
 
 void TerminalWriteCharacter(char c);
